* Use BLAS Level 3 matrix-matrix multiplications to compute pairwise distance in k-means.
* Remove ```runs``` related code completely, it will have no effect after this change.

cc @mengxr @jkbradley

Update:
* I have make performance test, but found the new version is slower than the old one(1.5 times).
Further more, I track the calling stack and found that the cost to construct the ```pointMatrix```, ```centerMatrix``` and ```distanceMatrix``` is expensive. I try to compute and cache ```pointMatrix``` and ```centerMatrix``` in advance, but it still can not get benefits. Looking forward others' comments.
* I also found ```gemm``` is slower than ```axpy``` in the ```mllib.linalg``` package. Consider the following code which is the abstract of ```KMeans``` distance computation scenarios of new and old version:
```Scala
    val n = 3000
    val random = new Random()
    val c1 = Matrices.zeros(n, n).asInstanceOf[DenseMatrix]
    val a1 = Matrices.randn(n, n, random).asInstanceOf[DenseMatrix]
    val b1 = Matrices.randn(n, n, random).asInstanceOf[DenseMatrix]

    val start1 = System.nanoTime()
    gemm(2.0, a1, b1, 2.0, c1)
    println("gemm elapsed time: = %.3f".format((System.nanoTime() - start1)/1e9) + " seconds.")

    val a2 = Vectors.dense(Array.fill(n)(random.nextDouble()))
    val aa = Array.fill(n)(a2)
    val b2 = Vectors.dense(Array.fill(n)(random.nextDouble()))
    val bb = Array.fill(n)(b2)

    val start2 = System.nanoTime()
    for (i <- 0 until n; j <- 0 until n) {
      axpy(1.0, bb(j), aa(j))
    }
    println("axpy elapsed time: = %.3f".format((System.nanoTime() - start2)/1e9) + " seconds.")
```
I got the performance result on my Mac:
```
gemm elapsed time: = 26.519 seconds.
axpy elapsed time: = 20.660 seconds.
```
It means we can not get benefits from BLAS Level 3 matrix-matrix multiplications to compute pairwise distance. I also found others' complains which is similar with this issue (https://github.com/xianyi/OpenBLAS/issues/528).
If I replace ```axpy(1.0, bb(j), aa(j))``` with ```dot(bb(j), aa(j))```, I got:
```
gemm elapsed time: = 28.938 seconds.
dot elapsed time: = 28.574 seconds.
```
Please correct me if I have some misunderstanding.