@mengxr 2016-01-15
Regarding your local performance test:

1. Make sure you installed optimized BLAS on your system and loaded correctly in JVM via netlib-java. The different should be significant at 3000x3000 (with or without multi-treading).
2. Your test of GEMM and AXPY is not equivalent. First of all, they are not using the same matrices for multiplication. Secondly, ` axpy(1.0, bb(j), aa(j))` should be ` axpy(1.0, bb(j), aa(i))`. Otherwise, you get some benefit from local caching.

Could you re-run the test? I will take a look at your implementation.