@yanboliang 2016-01-18
@mengxr I have a new and advanced implementation for this issue at #10806 , let's move the discussion there. I will close this PR now.